Refactor job page into sidebar components
